Create reference file
I decided to create a new file to act as a reference for why I made
the decisions I did when building this app. I have a feeling that I'll
be implementing a lot of code that I'm unfamiliar with and will need
to be able to explain my decisions clearly and thoroughly at the
project review. And I know there's no way I'll remember why I did
everything--hence this file!