Create application_controller
In the configure block, I enabled the use of sessions and set the
session secret by using a gem called Sysrandom to encode the user's
password. I also set other app configurations.

I created two helper methods: one to identify the current user and
another to identify if the user is logged in or not.